Update custom menu items patch
The previous version simply interpreted the value of the "x-canonical-type"
attribute on a menu item as a GType string and instantiated an instance of
that.  We need something more sophisticated now, because GMenuModel is used by
things that don't integrate with glib's type system. (In particular, the next
unity will be a consumer of the indicator services.)

This new patch adds an extension point to gtk. Libraries or programs can
register to implement this extension point and supply an object with the
GtkMenuItemFactory interface. Gtk will ask those factories to create menu items
whenever it stumbles on a "x-canonical-type" attribute.