css: replace rpx with vw/vh
rpx was a good idea but unfortunately, fractional scaling.

TODO: values need to be invalidated when screen is resized. Since this is really just
a stop-gap and Phosh will be on GTK4 sooner, I won't implement it. The spectacular hack
will be to simply wrap the value in calc() - that way, the value will be re-evaluated
when the screen is resized. Awful, I know.